+++ GamePriceFinder/Finders/NuuvemFinder.cs
@@ -103,7 +103,7 @@ namespace GamePriceFinder.Finders
                 game.Video = await YoutubeHandler.GetGameTrailer(string.Concat(name, TRAILER));
 #endif
 
-                var gamePrice = new GamePrices(game.GameId, ((int)StoresEnum.Nuuvem).ToString(), PriceHandler.ConvertPriceToDatabaseType(price, 3));
+                var gamePrice = new GamePrices(game.GameId, StoresEnum.Nuuvem.ToString(), PriceHandler.ConvertPriceToDatabaseType(price, 3));
                 var history = new History(game.GameId, StoresEnum.Nuuvem.ToString(), gamePrice.CurrentPrice, DateTimeOffset.Now.ToUnixTimeSeconds().ToString());
                 var genre = new Genre("Action");
                 return new DatabaseEntitiesHandler(game, gamePrice, history, genre);
